requirements.txt
django==5.0.1
django-ninja==1.1.0
django-extensions
django-debug-toolbar==4.3.0
django-filter==23.2  # https://django-filter.readthedocs.io/en/stable/index.html
django-cors-headers==4.1.0  # https://github.com/adamchainz/django-cors-headers
djangorestframework==3.14.0  # https://github.com/encode/django-rest-framework
djangorestframework-simplejwt==5.2.2  # https://django-rest-framework-simplejwt.readthedocs.io/en/latest/
drf-writable-nested==0.7.0  # https://pypi.org/project/drf-writable-nested/
# DRF-spectacular for api documentation
drf-spectacular==0.26.3  # https://github.com/tfranzel/drf-spectacular

graphene-django
elasticsearch_dsl==8.12.0
requests

# Code quality
black
ruff